Move startGame function to its own file.
I know this was not the best decision to make. startGame belongs in the
game module, but because I have a lot of problems when restarting the
game, I decided to move it.

The problems I was facing included not being able to create a new
instance of game inside the game module.